renamed a directory
